+++ options/__init__.py
@@ -1,8 +1,8 @@
 """Simple app to add configuration options to a Django project."""
+import django
+
 from options.constants import CONVERTER, FILE, FLOAT, INT, STR, TYPE_CHOICES
 from options.helpers import get_option_model, get_user_option_model
-
-default_app_config = "options.apps.ConfigurationsConfig"
 
 __all__ = [
     "get_option_model",
@@ -14,4 +14,7 @@ __all__ = [
     "TYPE_CHOICES",
     "CONVERTER",
 ]
-__version__ = "2.1.2"
+__version__ = "2.3.0"
+
+if django.VERSION < (3, 2):
+    default_app_config = "options.apps.ConfigurationsConfig"