+++ options/constants.py
@@ -1,3 +1,5 @@
+from typing import Dict, Type
+
 from django.utils.translation import gettext_lazy as _
 
 FLOAT, INT, STR, FILE = (0, 1, 2, 3)
@@ -7,4 +9,4 @@ TYPE_CHOICES = (
     (STR, _("String")),
     (FILE, _("File")),
 )
-CONVERTER = {INT: int, FLOAT: float, STR: str, FILE: str}
+CONVERTER: Dict[int, Type] = {INT: int, FLOAT: float, STR: str, FILE: str}